* Reformatted DatabaseConnector to work with both CLI and GUI * Fixed issue with logging in where account type wasn't being passed in correctly * Changed username to ID * Fixed phone number field being disabled when searching for an account
