+++ data/students.csv
@@ -1,4 +1,4 @@
 sID,name,password,sex,phone,tuition
 1,Derek Lopes,123,1,5309665355,0
 2,Derek Jeter,456,1,5555555555,100
-3,Bo Derek,789,0,2223334444,5000
+3,Bo Derek,789,0,1234567890,5000
